fix download bug

+++ ok/update/GithubMultiDownloader.py
@@ -31,7 +31,7 @@ class GithubMultiDownloader:
         self.fast_proxys = []
         random.shuffle(self.proxys)
         self.lock = threading.Lock()
-        self.num_parts = 5
+        self.num_parts = 6
         self.downloaded = 0
         self.start_time = 0
         self.size = 0
@@ -48,31 +48,35 @@ class GithubMultiDownloader:
             return url, proxy
 
     def download_part(self, part_size, start, end, file, giturl, last_proxy=None):
-        target_size = end - start + 1
-        if start >= end:
+        part_start = start
+        with self.lock:
+            if os.path.exists(file):
+                downloaded = os.path.getsize(file)
+                if downloaded == part_size:
+                    logger.info(f'File {file} already downloaded')
+                    return
+                elif downloaded > part_size:
+                    logger.warning(f'File size error {file} {downloaded} > {part_size}')
+                    os.remove(file)
+                else:
+                    part_start += downloaded
+                self.downloaded += part_size
+
+        target_size = end - part_start + 1
+
+        if target_size <= 0:
             logger.debug(f'end multi part downloading {part_size} {target_size} {file} {last_proxy}')
             with self.lock:
                 if last_proxy is not None:
                     self.fast_proxys.insert(0, last_proxy)
             return
 
-        if os.path.exists(file):
-            downloaded = os.path.getsize(file)
-            if downloaded == part_size:
-                logger.info(f'File {file} already downloaded')
-                return
-            elif downloaded > part_size:
-                logger.warning(f'File size error {file} {downloaded} != {target_size}')
-                os.remove(file)
-            else:
-                start = downloaded
-
-        headers = {'Range': f'bytes={start}-{end}'}
+        headers = {'Range': f'bytes={part_start}-{end}'}
         url, proxy = self.next_url(giturl)
         if url is None:
             logger.error(f'all proxies failed to download {file}')
             return
-        logger.debug(f'start multi part downloading {file} {url}')
+        logger.debug(f'start multi part downloading {part_size} {target_size} {target_size <= part_size} {file} {url}')
         start_time = time.time()
 
         try:
@@ -80,15 +84,16 @@ class GithubMultiDownloader:
             response_size = int(response.headers.get('Content-Length', 0))
 
             if response_size != target_size:
-                raise Exception(f'size mismatch: {response_size} != {target_size}')
+                response.close()
+                raise Exception(f'response_size mismatch: {response_size} != {target_size}')
             part_downloaded = 0
             last_chunk_time = time.time()
             with open(file, 'ab') as f:
                 for chunk in response.iter_content(chunk_size=1024):  # 1 KB chunks
                     with self.lock:
-                        if time.time() - last_chunk_time > 5:  # If more than 10 seconds have passed since the last chunk
+                        if time.time() - last_chunk_time > 5:  # If more than 5 seconds have passed since the last chunk
                             response.close()
-                            raise Exception(f'{proxy} Server is not responding with the chunk every 5 seconds')
+                            logger.error(f'{proxy} Server is not responding with the chunk every 5 seconds')
                             break
                         if self.exit_event.is_set():
                             return
@@ -96,7 +101,6 @@ class GithubMultiDownloader:
                             f.write(chunk)
                             self.downloaded += len(chunk)
                             part_downloaded += len(chunk)
-                            start += len(chunk)
                             percent = self.downloaded / self.size * 100
                             # Every 1%
                             communicate.download_update.emit(percent,
@@ -105,6 +109,8 @@ class GithubMultiDownloader:
                                                              False,
                                                              None
                                                              )
+                            if part_downloaded > target_size:
+                                logger.error(f'part_downloaded > target_size')
                         last_chunk_time = time.time()  # Update the time when the last chunk was received
                         if len(self.fast_proxys) > 0:
                             response.close()
@@ -132,7 +138,7 @@ class GithubMultiDownloader:
                 # The end byte is one less than the start byte of the next part
                 end = start + part_size - 1 if i < self.num_parts - 1 else self.size - 1
                 futures.append(
-                    executor.submit(self.download_part, part_size, start, end, file, release.get('url'), None))
+                    executor.submit(self.download_part, end - start + 1, start, end, file, release.get('url'), None))
 
         # Wait for all parts to finish downloading
         for future in futures:
@@ -143,13 +149,17 @@ class GithubMultiDownloader:
         with open(whole, 'wb') as fp:  # replace with your file path
             for i in range(self.num_parts):
                 file = self.get_part_name(dir, i, release)
+                if not os.path.exists(file):
+                    logger.error(f'file part does not exist: {file}')
+                    return False
                 with open(file, 'rb') as part_file:
                     fp.write(part_file.read())
                 os.remove(file)  # Delete the part file after combining
             combined_size = os.path.getsize(whole)
-            if combined_size != self.size:
-                logger.error(f'combined_size mismatch: {combined_size} != {self.size}')
-                return False
+        if combined_size != self.size:
+            logger.error(f'combined_size mismatch: {combined_size} != {self.size}')
+            os.remove(whole)
+            return False
         return True
 
     def get_part_name(self, update_dir, i, release):
